move connect-wallet wizard route under strategy
