+++ src/routes/strategies/[strategy]/(wizard)/+layout.ts
@@ -1,0 +1,9 @@
+export const ssr = false;
+
+export async function load() {
+	return {
+		skipNavbar: true,
+		skipFooter: true,
+		skipSideNav: true
+	};
+}
