+++ src/routes/strategies/[strategy]/(wizard)/connect-wallet/+layout.svelte
