+++ src/routes/strategies/[strategy]/(wizard)/connect-wallet/balance/+page.svelte
