+++ src/routes/strategies/[strategy]/(wizard)/connect-wallet/balance/+page.ts
@@ -1,15 +1,12 @@
-import type { ConnectWizardData } from '../+layout';
-import { get } from 'svelte/store';
-import { wizard } from '$lib/wizard/store';
 import { config } from '$lib/wallet/client';
 import { getAccount, getBalance } from '@wagmi/core';
 
-export async function load() {
+export async function load({ parent }) {
 	const { address } = getAccount(config) as { address: Address };
-	const { chainId } = get(wizard).data as ConnectWizardData;
+	const { chain } = await parent();
 
 	return {
-		nativeCurrency: await getBalance(config, { address, chainId }),
+		nativeCurrency: await getBalance(config, { address, chainId: chain.id }),
 		denominationToken: undefined
 		// TODO: fetch based on token address passed into wizard
 		// denominationToken: await fetchDenominationToken(chainId, onChainData, address)
