+++ src/routes/strategies/[strategy]/(wizard)/connect-wallet/introduction/+page.svelte
