+++ src/routes/strategies/[strategy]/(wizard)/connect-wallet/success/+page.svelte
